+++ src/main/java/de/sakurajin/evenbetterarcheology/api/item/BetterBrushItem.java
@@ -1,18 +1,27 @@
 package de.sakurajin.evenbetterarcheology.api.item;
 
+import com.google.gson.JsonArray;
+import com.google.gson.JsonObject;
+import de.sakurajin.evenbetterarcheology.registry.PatchouliBookGeneration;
+import de.sakurajin.sakuralib.arrp.v2.patchouli.JPatchouliEntry;
+import de.sakurajin.sakuralib.arrp.v2.patchouli.pages.JRecipePage;
+import de.sakurajin.sakuralib.arrp.v2.patchouli.pages.JTextPage;
 import de.sakurajin.sakuralib.datagen.v1.DatagenModContainer;
 import de.sakurajin.sakuralib.datagen.v1.DataGenerateable;
 import io.wispforest.owo.itemgroup.OwoItemSettings;
+import io.wispforest.owo.text.TranslationContext;
 import net.devtech.arrp.json.recipe.*;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.BrushableBlock;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BrushableBlockEntity;
+import net.minecraft.client.resource.language.TranslationStorage;
 import net.minecraft.entity.EquipmentSlot;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.*;
+import net.minecraft.registry.Registries;
 import net.minecraft.sound.SoundCategory;
 import net.minecraft.sound.SoundEvent;
 import net.minecraft.sound.SoundEvents;
@@ -24,11 +33,15 @@ import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.hit.HitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.List;
+import java.util.Map;
 
 /**
  * This is an extension of the BrushItem for Brushes that are made from better materials.
  * It is recommended to use the Builder to create a new instance of this class.
- *
+ * <p>
  * To generate the data using the data generator, a material has to be provided.
  * This way the data generation system can generate the crafting recepie.
  * The material has to be an Item that is registered in the game.
@@ -38,8 +51,8 @@ import net.minecraft.world.World;
  * @see Builder
  */
 public class BetterBrushItem extends BrushItem implements DataGenerateable {
-    private final float brushingSpeed;
-    private final Item material;
+    private final float  brushingSpeed;
+    private final Item   material;
     private final String materialName;
 
 
@@ -49,34 +62,46 @@ public class BetterBrushItem extends BrushItem implements DataGenerateable {
         this.material = material;
 
         String materialTranslationKey = material.getTranslationKey();
-        var materialArray = materialTranslationKey.split("\\.");
-        this.materialName = materialArray[materialArray.length-1];
+        var    materialArray          = materialTranslationKey.split("\\.");
+        this.materialName = materialArray[materialArray.length - 1];
     }
 
     @Override
     public ItemConvertible generateData(DatagenModContainer container, String identifier) {
         container.addTag("c:items/brushitems", identifier);
         container.generateItemModel(identifier, "minecraft:item/brush", identifier);
-        if(material != null) {
+
+        JTextPage mainPage = JTextPage.create("patchouli_book.even_better_archeology.better_archeology_guide.brush." + identifier);
+
+        JPatchouliEntry brushEntry = JPatchouliEntry
+            .create("item.evenbetterarcheology."+identifier, PatchouliBookGeneration.BRUSHES_CATEGORY_ID.toString(), container.getStringID(identifier))
+            .addPage(mainPage);
+
+        if (material != null) {
+            Identifier recipeID = container.getSimpleID("crafting_" + identifier + "_" + materialName);
             container.RESOURCE_PACK.addRecipe(
-                    new Identifier(identifier, "crafting_" + identifier + "_" + materialName),
-                    JRecipe.shaped(
-                            JPattern.pattern("x", "y", "z"),
-                            JKeys.keys()
-                                    .key("x", JIngredient.ingredient().item(material))
-                                    .key("y", JIngredient.ingredient().item(Items.STICK))
-                                    .key("z", JIngredient.ingredient().item(Items.FEATHER)),
-                            JResult.item(this)
-                    )
+                recipeID,
+                JRecipe.shaped(
+                    JPattern.pattern("x", "y", "z"),
+                    JKeys.keys()
+                         .key("x", JIngredient.ingredient().item(material))
+                         .key("y", JIngredient.ingredient().item(Items.STICK))
+                         .key("z", JIngredient.ingredient().item(Items.FEATHER)),
+                    JResult.item(this)
+                )
             );
-        }else{
+
+            brushEntry = brushEntry.addPage(JRecipePage.create(recipeID.toString()));
+        } else {
             container.LOGGER.warn("No material provided for the BetterBrushItem {}. The crafting recipe will not be generated.", identifier);
         }
+
+        container.registerPatchouliEntry(PatchouliBookGeneration.BOOK_NAME, brushEntry);
 
         return this;
     }
 
-    public float getBrushingSpeed(){
+    public float getBrushingSpeed() {
         return brushingSpeed;
     }
 
@@ -85,8 +110,8 @@ public class BetterBrushItem extends BrushItem implements DataGenerateable {
             HitResult hitResult = this.getHitResult(user);
             if (hitResult instanceof BlockHitResult blockHitResult && hitResult.getType() == HitResult.Type.BLOCK) {
                 //actually brush the block
-                if ((this.getMaxUseTime(stack) - remainingUseTicks + 1) % brushingSpeed == brushingSpeed/2) {
-                    BlockPos blockPos = blockHitResult.getBlockPos();
+                if ((this.getMaxUseTime(stack) - remainingUseTicks + 1) % brushingSpeed == brushingSpeed / 2) {
+                    BlockPos   blockPos   = blockHitResult.getBlockPos();
                     BlockState blockState = world.getBlockState(blockPos);
 
                     // spawn the particles
@@ -125,7 +150,7 @@ public class BetterBrushItem extends BrushItem implements DataGenerateable {
         user.stopUsingItem();
     }
 
-    public static Builder Builder(DatagenModContainer container){
+    public static Builder Builder(DatagenModContainer container) {
         return new Builder(container);
     }
 
@@ -133,9 +158,9 @@ public class BetterBrushItem extends BrushItem implements DataGenerateable {
      * A Builder for the BetterBrushItem.
      * It is recommended to use this Builder to create a new instance of the BetterBrushItem.
      * The default settings are created using the DatagenModContainer.
-     *
+     * <p>
      * The following properties can be set using the Builder:
-     *
+     * <p>
      * * brushingSpeed (default: 1.0f) - lower values mean faster brushing
      * * maxDamage (default: container.settings()) - the maximum durability of the brush
      * * maxCount (default: container.settings()) - the maximum stack size of the brush
@@ -145,41 +170,41 @@ public class BetterBrushItem extends BrushItem implements DataGenerateable {
      * @see BetterBrushItem
      * @see DatagenModContainer
      */
-    public static class Builder{
-        private float brushingSpeed = 1.0f;
+    public static class Builder {
+        private       float           brushingSpeed = 1.0f;
         private final OwoItemSettings settings;
-        private Item material = null;
+        private       Item            material      = null;
 
         public Builder(DatagenModContainer container) {
             this.settings = container.settings();
         }
 
-        public Builder setBrushingSpeed(float pBrushingSpeed){
+        public Builder setBrushingSpeed(float pBrushingSpeed) {
             brushingSpeed = pBrushingSpeed;
             return this;
         }
 
-        public Builder setMaxDamage(int maxDamage){
+        public Builder setMaxDamage(int maxDamage) {
             settings.maxDamage(maxDamage);
             return this;
         }
 
-        public Builder setMaxCount(int maxCount){
+        public Builder setMaxCount(int maxCount) {
             settings.maxCount(maxCount);
             return this;
         }
 
-        public Builder setRarity(Rarity rarity){
+        public Builder setRarity(Rarity rarity) {
             settings.rarity(rarity);
             return this;
         }
 
-        public Builder setMaterial(Item material){
+        public Builder setMaterial(Item material) {
             this.material = material;
             return this;
         }
 
-        public BetterBrushItem build(){
+        public BetterBrushItem build() {
             return new BetterBrushItem(settings, brushingSpeed, material);
         }